+++ scripts/hook.php
@@ -6,6 +6,7 @@
 
 define('ROOT_PATH', dirname(dirname(__FILE__)) . '/');
 define('NO_SLIM', 1);
+define('NO_SESSION', 1);
 
 require_once ROOT_PATH . 'lib/bootstrap.php';
 date_default_timezone_set('Europe/Berlin');